+++ test/amd/test_dry.py
@@ -3,8 +3,9 @@ import json
 import pytest
 
 from eye_extractor.amd.dry import DRY_PAT, DRY_AMD_PAT, extract_dryamd_severity
+from eye_extractor.amd.ga import GeoAtrophy
 from eye_extractor.headers import Headers
-from eye_extractor.output.amd import build_dryamd_severity
+from eye_extractor.output.amd import build_dryamd_severity, augment_dryamd_severity
 
 
 @pytest.mark.parametrize('pat, text, exp', [
@@ -18,9 +19,10 @@ def test_dry_patterns(pat, text, exp):
 
 
 @pytest.mark.parametrize('text, headers, exp_dryamd_severity_re, exp_dryamd_severity_le, exp_dryamd_severity_unk', [
-    ('', {'MACULA': 'dry'}, 'UNKNOWN', 'UNKNOWN', 'UNSPECIFIED'),
-    ('', {'ASSESSMENT': 'atrophy od'}, 'UNSPECIFIED', 'UNKNOWN', 'UNKNOWN'),
-    ('atrophic armd', None, 'UNKNOWN', 'UNKNOWN', 'UNSPECIFIED'),
+    ('', {'MACULA': 'dry'}, 'UNKNOWN', 'UNKNOWN', 'YES'),
+    ('', {'ASSESSMENT': 'atrophy od'}, 'YES', 'UNKNOWN', 'UNKNOWN'),
+    ('atrophic armd', None, 'UNKNOWN', 'UNKNOWN', 'YES'),
+    ('OD: non-exudative senile AMD', None, 'YES', 'UNKNOWN', 'UNKNOWN'),
 ])
 def test_dry_extract_build(text, headers, exp_dryamd_severity_re, exp_dryamd_severity_le, exp_dryamd_severity_unk):
     pre_json = extract_dryamd_severity(text, headers=Headers(headers))
@@ -29,3 +31,10 @@ def test_dry_extract_build(text, headers, exp_dryamd_severity_re, exp_dryamd_sev
     assert result['dryamd_severity_re'] == exp_dryamd_severity_re
     assert result['dryamd_severity_le'] == exp_dryamd_severity_le
     assert result['dryamd_severity_unk'] == exp_dryamd_severity_unk
+
+
+def test_dry_augment():
+    ga_result = {'geoatrophy_re': GeoAtrophy.YES}
+    dry_result = {'dryamd_severity_re': 'UNKNOWN'}
+    res = augment_dryamd_severity(dry_result, ga_result=ga_result)
+    assert res['dryamd_severity_re'] == 'YES'