+++ test/amd/test_fluid.py
@@ -87,6 +87,10 @@ def test_fluid_to_column(data, exp_fluid_amd_re, exp_fluid_amd_le):
 @pytest.mark.parametrize('text, headers, exp_fluid_re, exp_fluid_le, exp_fluid_unk', [
     ('', {'MACULA': 'subretinal fluid od'}, 'SUBRETINAL FLUID', 'UNKNOWN', 'UNKNOWN'),
     ('', {'MACULA': 'with fluid and exudates'}, 'UNKNOWN', 'UNKNOWN', 'FLUID'),
+    ('', {'MACULA': 'large area of edema OD'}, 'FLUID', 'UNKNOWN', 'UNKNOWN'),
+    ('corneal fluid', {}, 'UNKNOWN', 'UNKNOWN', 'UNKNOWN'),
+    ('corneal edema', {}, 'UNKNOWN', 'UNKNOWN', 'UNKNOWN'),
+    ('macular edema os', {}, 'UNKNOWN', 'FLUID', 'UNKNOWN'),
 ])
 def test_fluid_extract_build(text, headers, exp_fluid_re, exp_fluid_le, exp_fluid_unk, ):
     pre_json = extract_fluid(text, headers=Headers(headers))